Add files via upload
In this project, I have contrasted linear regression and classification by plotting graphs. Linear regression is not always sufficient for solving a problem because many problems have just a finite number of outputs. Here's where classification comes in. Classification gives us a more accurate answer compared to the variety of outputs that are possible with linear regression.