Add files via upload
In this project, I have taken an example as to how we can determine whether the coffee beans have been roasted well, or not. There are two parameters in this example: duration and temperature. We create many hidden layers of neurons and use activation function also called the sigmoid function. This is also called a dense layer function.