Add files via upload
In this project, I have taken an example of determining housing prices based on the size of the house. We use linear regression in this. The cost function tells us how well the model is doing so we can try to do better. The parameters w and b are the variables you can adjust during training in order to improve the model. The graph will be different for different values of w and b. Linear regression will find out values of w and b such that the cost function will give a value as small as possible.