Add files via upload
In this project, I learnt about feature engineering. Feature engineering is the process of using intuition to design new features by transforming or combining original features. The model can choose any number of parameters depending on whether the data shows the frontage or the depth or the area of the plot for predicting the price of the house in the example.