Add files via upload
In this project, I have updated gradient descent for logistic regression. To fit the parameters of a logistic regression model, we have to find the values of parameters w and b that minimize the cost function. I have used the average value to minimize the cost function. Logistic and linear regression have similar concepts but different equations and purposes.