Add files via upload
In this project, I have used linear regression. Linear regression predicts a number of infinitely many possible outputs. The goal of this project is to fit a straight line on the graph covering the common data. The data set used to train the model is called a training set. This particular model in the project us called linear regression with one variable or univariable linear regression.