Add files via upload
In this project, I have used logistic loss function. Loss is the measure of a single example to its target value while the cost is the measure of losses over the training sets. The defining feature of the loss function is that it uses two separate curves. This is suitable for classification problems.