Add files via upload
In this project, I have explored neurons and layers using packages like numpy, tensorflow and keras. A layer is a grouping of neurons which takes as input the same or similar features and that in turn outputs a few numbers together. A layer can have multiple neurons or  it can also have a single neuron.