Add files via upload
In this project, I have just implemented the basic functions from the numpy library in python. Numpy allocates memories using vectors. Indexing means referring to an element of an array by its position in the array. Slicing means getting a subset of elements of an array based on its indices. These are just 2 examples of the vast functions present in the numpy library.