Add files via upload
In this small project, I have tried to give a situation where overfitting might occur. If the training set has fit the data extremely well, it is called overfitting or it is said to have high variance. To address overfitting, we can collect more training data and we can select features to include or exclude from the model.