Add files via upload
In this project, I have implemented regularization. Regularization is a way to gently reduce the effects of some features without directly implementing them. It is less likely to overfit and gives a simpler model. Lambda is called the regularization parameter. By scaling all the terms the same way, it becomes easier to choose an appropriate value for lambda.