Add files via upload
In this project, I have built a simple neural network using the coffee roasting example with the help of a library called tensorflow. 2 features called duration and temperature are very important in this example. I created a normalization layer and 'adapted' the data. This learns the mean and variance and saves the values internally. Then I normalized the data. We use 2 key functions called model.fit and model.compile .